requirements.txt
python-telegram-bot==20.7
python-bitcoinrpc==1.0
peewee==3.17.0
psycopg2-binary==2.9.9
requests==2.31.0
python-dotenv==1.0.0
cryptography==41.0.8
